refactor(anthropic): extract key sanitizer to core and dereference $ref
P3-5 from the review. Move the schema property-key sanitization mechanism into
`@composio/core` and have the provider supply only the Anthropic constraint, then
dereference `$ref` before sanitizing so the previously-un-restorable `$ref`/`$defs`
positions become ordinary restorable value positions.

Core (`utils/schemaPropertyKeys.ts`, exported from the package root):
- `sanitizeSchemaPropertyKeys(schema, policy)`, `restoreOriginalKeys`,
  `mappingHasRenames`, and the `KeyMapping` / `KeySanitizationPolicy` types.
- The traversal, collision handling, schema-shaped reverse mapping, prototype
  safety and depth cap are provider-agnostic; the key constraint is injected as a
  policy. Reuses core's `isPlainObject`.

Anthropic provider:
- `key-policy.ts` holds the Anthropic constraint (`^[a-zA-Z0-9_.-]{1,64}$`, the
  `$`/`@` aliases, length truncation, collision resolution) as a
  `KeySanitizationPolicy`. It only imports the policy *type* from core, so it stays
  runtime-pure.
- `sanitize-keys.ts` is now a thin binding that re-exports core's restore side and
  a policy-bound `sanitizeSchemaPropertyKeys`, keeping the import path stable for
  `index.ts` and the tests.
- `wrapTool` dereferences internal `$ref`/`$defs` (lenient `sentinel` mode, so a
  dangling ref degrades to a permissive schema instead of throwing) before
  sanitizing. Keys reachable only through a reference are now both sanitized and
  restored; only genuinely dynamic positions (`additionalProperties`,
  `patternProperties`, `contains`) remain rewrite-only.

Adds a provider test proving a `$ref`-nested illegal key round-trips end to end.
