+++ ts/packages/core/src/index.ts
@@ -16,6 +16,12 @@ export type {
 } from './utils/jsonSchema';
 export { getExtensionFromMimeType } from './utils/mime';
 export { normalizeToolArguments } from './utils/toolArguments';
+export {
+  sanitizeSchemaPropertyKeys,
+  restoreOriginalKeys,
+  mappingHasRenames,
+} from './utils/schemaPropertyKeys';
+export type { KeyMapping, KeySanitizationPolicy } from './utils/schemaPropertyKeys';
 export { AuthScheme } from './models/AuthScheme';
 export { MCP } from './models/MCP';
 export { RemoteFile } from './models/RemoteFile';
